chore: move schemas to new package
